Tiny fix

+++ core/hash.h
@@ -1,6 +1,6 @@
 #ifndef P2P_TECHDEMO_CORE_HASH_H
 #define P2P_TECHDEMO_CORE_HASH_H
 
-size_t hash32();
+uint32_t hash32();
 
 #endif // P2P_TECHDEMO_CORE_HASH_H
